[CRES-32] refactor: 디데이 처리 함수 utils 폴더로 분리해서 사용 #20
